Update infrared-openstack.sh 13 script
Remove the PREFIX option that I don't think we ever got working correctly. Import
the ENVIRONMENT_TEMPLATE variable that we use in the 16 script for future considerations.

Bump the base OS image and puddle target to match current phase 2 tests for OSP13.

Reduce CPU and memory footprint to match recent changes in 16 script.